addrs: NewDefaultProviderConfig function
This is a helper function for creating a "default" (unaliased) provider
configuration address for a given provider type.

This is not doing anything that a caller couldn't do directly itself, but
it's provided so that callers can express their intent more clearly
through this more-specific function name and associated documentation.